pyproject.toml
[tool.poetry]
name = "dbmigrator"
version = "0.1.0"
description = ""
authors = ["Victor Misael <misael1805@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
mysql-connector-python = "^8.3.0"
psycopg2 = "^2.9.9"
load-dotenv = "^0.1.0"
flask = "^2.1.0"
flask-socketio = "^5.1.0"
eventlet = "^0.35.0"  # necessário para o Flask-SocketIO


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
